gen-ai-course/05_mcp/04_mcp_enterprise_project/pyproject.toml
[project]
name = "enterprise-mcp-server"
version = "0.1.0"
description = "An enterprise-ready MCP server scaffold using uv and Python."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "mcp>=1.8.0",
  "opentelemetry-api>=1.28.0",
  "opentelemetry-exporter-otlp-proto-http>=1.28.0",
  "opentelemetry-instrumentation-starlette>=0.49b0",
  "opentelemetry-sdk>=1.28.0",
  "pyjwt[crypto]>=2.10.0",
  "python-dotenv>=1.0.1",
  "uvicorn>=0.30.0",
]

[dependency-groups]
dev = [
  "pytest>=8.2.0",
  "ruff>=0.5.0",
]

[project.scripts]
enterprise-mcp-server = "enterprise_mcp.server:main"
enterprise-mcp-http = "enterprise_mcp.http_app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/enterprise_mcp"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
